+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 Version 1.9-SNAPSHOT
 -------------
 
+ADDED:
 
+- new method in class SerializedObjectExtensions for convert an object to a base64 encoded String object
+- new method in class Serializer for convert an object to a base64 encoded String object
+- new method in class SerializedObjectExtensions for convert a base64 encoded String object back to an object
+- new method in class Serializer for convert a base64 encoded String object back to an object
 
 Version 1.8
 -------------